+++ Client/src/scss/global/_variables.scss
@@ -9,6 +9,8 @@ $AdminNavBarMenuColor: #ab7d45;
 $AdminNavBarMenuHoverColor: #601c77;
 $SeparatorColor: #555;
 $InkColor: #373a3d;
+$ErrorColor: #b5553e;
+$LightFontColor: white;
 
 // Layout
 $AdminSidebarWidth: 300px;
